dojo-facturas/requirements.txt
fastapi
uvicorn
pydantic
anthropic
python-dotenv
python-multipart
psycopg2-binary
